feat se crea dto y mapa

+++ src/main/java/com/example/proyectoIntegradorIyR/repositorios/IUsuarioRepositorio.java
@@ -1,12 +1,11 @@
 package com.example.proyectoIntegradorIyR.repositorios;
 
 import com.example.proyectoIntegradorIyR.modelos.Usuario;
-import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface IUsuarioRepositorio extends JpaRepository<Usuario, Long> {
+public interface IUsuarioRepositorio extends JpaRepository<Usuario, Integer> {
 
     Optional<Usuario> findByCorreo(String correo);
